more type stuff

+++ spuce/dsp_functions/partial_convolve.h
@@ -6,8 +6,8 @@ namespace spuce {
 // \brief Template partial convolve function
 //! \author Tony Kirke,  Copyright(c) 2001
 //! \ingroup template_array_functions misc
-template <class T> std::vector<T> partial_convolve(const std::vector<T>& x, const std::vector<T>& y, int N, int M) {
-  int i, j;
+template <class T> std::vector<T> partial_convolve(const std::vector<T>& x, const std::vector<T>& y, size_t N, size_t M) {
+  size_t i, j;
   auto L = M + N - 1;
   std::vector<T> c(L);
   for (i = 0; i < L; i++) {
