+++ spuce/filters/chebyshev_iir.cpp
@@ -29,13 +29,13 @@ void chebyshev_iir(iir_coeff& filt, float_type fcd, float_type ripple = 3.0) {
 //! Calculate poles (chebyshev)
 void chebyshev_s(iir_coeff& filt, float_type wp, float_type epi, size_t n) {
   auto l = 1;
-  int n2 = (n + 1) / 2;
+  size_t n2 = (n + 1) / 2;
   float_type arg;
   float_type x = 1 / epi;
   float_type lambda = pow(x*(1.0 + sqrt(1.0 + epi*epi)),1.0/n);
   float_type sm = 0.5*((1.0/lambda) - lambda);
   float_type cm = 0.5*((1.0/lambda) + lambda);
-  for (int j = 0; j < n2; j++) {
+  for (size_t j = 0; j < n2; j++) {
     arg = 0.5 * M_PI * (2*l-1) / ((float_type)(n));
     if (filt.get_type()==filter_type::low) {
         filt.set_pole(-wp * std::complex<float_type>(sm * sin(arg), cm * cos(arg)), n2-1-j);
